refactor(meta): keep lore meta command in shell/cmds; scanning logic in pkg/meta
The command adapter belongs in the shell (plumbing); the frontmatter-scanning
business logic belongs in a dedicated package. Because pkg/openlore imports
pkg/shell/cmds, cmds cannot import openlore back, so the business logic lives in
a new neutral pkg/meta (depends only on pkg/vfs and pkg/okf).

- pkg/meta: Scan/Record/Extender (moved from openlore/meta.go)
- pkg/shell/cmds/lore_meta.go: cmdLoreMeta adapter + registration
- plugin meta extenders now injected per session (CmdContext.MetaExtenders),
  mirroring Docsets/PublishTargets, replacing the openlore global registry
- okf plugin returns []meta.Extender
