+++ pkg/openlore/okf_plugin.go
@@ -7,6 +7,7 @@ import (
 	"path"
 
 	"github.com/aakarim/go-openlore/internal/config"
+	"github.com/aakarim/go-openlore/pkg/meta"
 	"github.com/aakarim/go-openlore/pkg/okf"
 	"github.com/aakarim/go-openlore/pkg/vfs"
 )
@@ -134,8 +135,8 @@ func (p *okfPlugin) WriteMiddleware() []WriteMiddleware {
 //
 //	"okf": {"valid": true}                       // conformant
 //	"okf": {"valid": false, "error": "<reason>"} // non-conformant
-func (p *okfPlugin) MetaExtenders() []MetaExtender {
-	return []MetaExtender{
+func (p *okfPlugin) MetaExtenders() []meta.Extender {
+	return []meta.Extender{
 		func(absPath string, content []byte, _ map[string]any) map[string]any {
 			oc := p.resolve(absPath)
 			if oc == nil || !matchesOKFPatterns(absPath, oc.Patterns) {
