+++ pkg/openlore/server.go
@@ -20,6 +20,7 @@ import (
 	"github.com/aakarim/go-openlore/internal/metrics"
 	"github.com/aakarim/go-openlore/internal/passkeys"
 	"github.com/aakarim/go-openlore/internal/skills"
+	"github.com/aakarim/go-openlore/pkg/meta"
 	"github.com/aakarim/go-openlore/pkg/shell"
 	"github.com/aakarim/go-openlore/pkg/shell/cmds"
 	"github.com/aakarim/go-openlore/pkg/vfs"
@@ -80,6 +81,10 @@ type Server struct {
 	// front of Stat/ReadDir/ReadFile in registration order. Empty in core
 	// go-openlore; when empty no read wrapper is installed (zero read overhead).
 	readMW []ReadMiddleware
+
+	// metaExtenders are the `lore meta` extenders contributed by plugins,
+	// installed per session in buildSessionShell.
+	metaExtenders []meta.Extender
 
 	// postCommitMW is the post-commit middleware contributed by plugins, run at
 	// the applier after a durable commit (feed emit, post_write hooks) in
@@ -658,9 +663,7 @@ func (s *Server) registerPlugin(p any) {
 		}
 	}
 	if mp, ok := p.(MetaExtenderProvider); ok {
-		for _, e := range mp.MetaExtenders() {
-			registerMetaExtender(e)
-		}
+		s.metaExtenders = append(s.metaExtenders, mp.MetaExtenders()...)
 	}
 	// Record the plugin's identity + version in the boot logs. Logged per
 	// registration so it captures plugins registered after NewServer (e.g. the
@@ -828,6 +831,7 @@ func (s *Server) buildSessionShell(id Identity) *shell.Shell {
 	// `publish`. Computed once here, where the access authority lives.
 	sh.SetDocsets(s.sessionDocsets(id))
 	sh.SetPublishTargets(s.sessionPublishTargets(id))
+	sh.SetMetaExtenders(s.metaExtenders)
 
 	// Set identity info as environment variables
 	if id.IdentityName != "" {
